disable changing allowed users if not allowed
It's entirely possible that this will need to move to actually using the ACL system, especially if a new role is called for,
but for now, this gets the functionality in fast enough. Any more complex solution just uses the same basic logic:
Check allowed roles and check ownership, just wrapped by more infrastructure.

Fix #7